add pseudocount to bdgcmp
MACS2 has a bug in a specific script that creates a lambda assertion error when
there is a 0 count in the region.

The recommendation is to add a small pseudocount to the  bedgraph prior to running ppois.

an intermediate bedgraph is created, p01_{...}, for treat and control where 0.1 has been added.
Once bdgcmp is complete, this file is removed to reclaim space.